Get rid of a shitload of type casts for formatted strings and use %zu in some places
git-svn-id: http://svn.ratbox.org/svnroot/ircd-ratbox/trunk@24894 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ src/sslproc.c
@@ -669,8 +669,8 @@ start_zlib_session(struct Client *server)
 	if(len > READBUF_SIZE)
 	{
 		rb_free(buf);
-		sendto_realops_flags(UMODE_ALL, L_ALL, "ssld - attempted to pass message of %ld len, max len %d, giving up", (long)len, READBUF_SIZE);
-		ilog(L_MAIN, "ssld - attempted to pass message of %ld len, max len %d, giving up", (long)len, READBUF_SIZE);
+		sendto_realops_flags(UMODE_ALL, L_ALL, "ssld - attempted to pass message of %zd len, max len %d, giving up", len, READBUF_SIZE);
+		ilog(L_MAIN, "ssld - attempted to pass message of %zd len, max len %d, giving up", len, READBUF_SIZE);
 		exit_client(server, server, server, "ssld readbuf exceeded");
 		return;
 	}
